+++ hypervisor/cloudhypervisor/start.go
@@ -7,7 +7,6 @@ import (
 	"syscall"
 
 	"github.com/projecteru2/core/log"
-	"golang.org/x/sys/unix"
 
 	"github.com/cocoonstack/cocoon/hypervisor"
 )
@@ -37,7 +36,6 @@ func (ch *CloudHypervisor) launchProcess(ctx context.Context, rec *hypervisor.VM
 		defer logFile.Close() //nolint:errcheck
 	}
 
-	raiseVMMRlimits(ctx)
 	// shell out: the cloud-hypervisor binary is the authoritative VMM.
 	cmd := exec.Command(ch.conf.CHBinary, args...) //nolint:gosec
 	// Setpgid so CH survives if this process exits.
@@ -60,19 +58,4 @@ func (ch *CloudHypervisor) launchProcess(ctx context.Context, rec *hypervisor.VM
 	// Daemon mode: parent must wait() or zombie blocks IsProcessAlive on stop/delete.
 	go cmd.Wait() //nolint:errcheck
 	return pid, nil
-}
-
-func raiseVMMRlimits(ctx context.Context) {
-	logger := log.WithFunc("cloudhypervisor.raiseVMMRlimits")
-	inf := unix.Rlimit{Cur: unix.RLIM_INFINITY, Max: unix.RLIM_INFINITY}
-	if err := unix.Setrlimit(unix.RLIMIT_MEMLOCK, &inf); err != nil {
-		logger.Warnf(ctx, "raise RLIMIT_MEMLOCK: %v", err)
-	}
-	var nofile unix.Rlimit
-	if err := unix.Getrlimit(unix.RLIMIT_NOFILE, &nofile); err == nil && nofile.Cur < nofile.Max {
-		nofile.Cur = nofile.Max
-		if err := unix.Setrlimit(unix.RLIMIT_NOFILE, &nofile); err != nil {
-			logger.Warnf(ctx, "raise RLIMIT_NOFILE: %v", err)
-		}
-	}
 }
